+++ src/controllers/ReportController.php
@@ -5,6 +5,9 @@ namespace digitalpulsebe\utmtracker\controllers;
 use craft\web\Controller;
 use yii\console\ExitCode;
 
+/**
+ * @deprecated since 3.1.0
+ */
 class ReportController extends Controller
 {
     public $enableCsrfValidation = false;